+++ tests/toisto/model/quiz/test_progress.py
@@ -106,7 +106,7 @@ class ProgressTest(ToistoTestCase):
     def test_next_quiz_is_quiz_with_progress(self):
         """Test that the next quiz is one the user has seen before if possible."""
         concepts = [create_concept(f"id{index}", dict(fi=f"fi{index}", nl=f"nl{index}")) for index in range(5)]
-        quizzes = Quizzes(quiz for quiz in create_quizzes("fi", "nl", *concepts) if quiz.quiz_types == ("listen",))
+        quizzes = Quizzes(quiz for quiz in create_quizzes("fi", "nl", *concepts) if quiz.quiz_types == ("dictate",))
         random_quiz = next(iter(quizzes))
         self.progress.mark_correct_answer(random_quiz)
         self.progress.get_retention(random_quiz).skip_until = None